requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
requests==2.32.0
apscheduler==3.10.4
pydantic==2.9.0
psycopg2-binary==2.9.9
